zsh,vim: Fix keymap problems in zshrc and improve vimtex
zshrc: There was some problem with the order in which keybindings were
assigned. Notably, vim mode was set after setting up some custom
bindings, which apparently interfered with the correct resolution of
names via terminfo.

Moreover, there was a bug that we only set up the keymap for the viins
keymap. Now we also bind for the vicmd.
Known Bug: The Backspace key is, as returned by terminfo, apparently
still wrong, though I have no idea why. Therefore the binding for it in
vicmd mode is useless.

vim: improve the way we use vimtex. Also, set up some maps for dealing
with buffers instead of tabs